rename: app, main, globalstyle -> app  폴더로 이동
