+++ src/app/main.tsx
@@ -1,10 +1,10 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import { worker } from './mocks/browser';
+import { worker } from '../mocks/browser';
 
 import './globals.scss';
-import App from '@/App';
+import App from '@/app/App';
 
 if (process.env.NODE_ENV === 'development') {
   worker.start();
