+++ src/app/styles/global.d.ts
@@ -1,4 +1,4 @@
-declare module "*.scss" {
+declare module '*.scss' {
   const content: { [className: string]: string };
   export default content;
 }
